% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE.import.R
\name{ATE.import}
\alias{ATE.import}
\title{Imports Angstrom Engineering Thermal Evaporator Log Data}
\usage{
ATE.import(filename)
}
\arguments{
\item{filename}{CSV filename including path for the ATE log file}
}
\value{
data frame with around 50 variables and rows that represent time;
      the variables include Date, Time, SubstrateShutterOpen, ChamberPressure
      and many other parameters.
}
\description{
Imports Angstrom Engineering Thermal Evaporator Log Data
}
\examples{
fileName = ATE.sampleFiles("csv")[1]
d = ATE.import(fileName)
head(d)

}
\author{
Thomas Gredig
}
