% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE.status.R
\name{ATE.status}
\alias{ATE.status}
\title{ATE Recipe Status}
\usage{
ATE.status(filename)
}
\arguments{
\item{filename}{path and filename of XML Status file from Angstrom Thermal Evaporator}
}
\value{
data frame with description steps, start and end times
}
\description{
Reads an XML Status file from the ATE thermal evaporator and returns
the procedural timeline.
}
\examples{
fileName <- ATE.sampleFiles('_Status')
ATE.status(fileName)

}
\author{
Thomas Gredig
}
